+++ firebase-ai/src/main/kotlin/com/google/firebase/ai/type/AudioHelper.kt
@@ -140,7 +140,6 @@ internal class AudioHelper(
    * Returns an empty flow if this [AudioHelper] has been [released][release].
    */
   fun listenToRecording(): Flow<ByteArray> {
-    println("Released: $released")
     if (released) return emptyFlow()
     resumeRecording()