+++ firebase-ai/src/main/kotlin/com/google/firebase/ai/type/LiveSession.kt
@@ -37,7 +37,6 @@ import java.util.concurrent.ConcurrentLinkedQueue
 import java.util.concurrent.atomic.AtomicBoolean
 import kotlin.coroutines.CoroutineContext
 import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.cancel
 import kotlinx.coroutines.channels.Channel.Factory.UNLIMITED
 import kotlinx.coroutines.flow.Flow
@@ -93,6 +92,11 @@ internal constructor(
    * @param functionCallHandler A callback function that is invoked whenever the model receives a
    * function call. The [FunctionResponsePart] that the callback function returns will be
    * automatically sent to the model.
+   *
+   * @param enableInterruptions Boolean to enable user to interrupt the model. Setting this variable
+   * would allow the user to talk while the model is responding.
+   *
+   * **WARNING**: User interruption might not work reliably across all devices.
    */
   @RequiresPermission(RECORD_AUDIO)
   public suspend fun startAudioConversation(
@@ -175,11 +179,9 @@ internal constructor(
             response
               .getOrNull()
               ?.let {
-                val x = JSON.decodeFromString<InternalLiveServerMessage>(
+                JSON.decodeFromString<InternalLiveServerMessage>(
                   it.readBytes().toString(Charsets.UTF_8)
                 )
-                println(x)
-                x
               }
               ?.let { emit(it.toPublic()) }
             yield()
@@ -233,7 +235,6 @@ internal constructor(
           BidiGenerateContentToolResponseSetup(functionList.map { it.toInternalFunctionCall() })
             .toInternal()
         )
-      println("Sending function response $jsonString")
       session.send(Frame.Text(jsonString))
     }
   }
@@ -253,7 +254,6 @@ internal constructor(
         Json.encodeToString(
           BidiGenerateContentRealtimeInputSetup(mediaChunks.map { (it.toInternal()) }).toInternal()
         )
-      println("Sending $jsonString")
       session.send(Frame.Text(jsonString))
     }
   }
@@ -310,7 +310,7 @@ internal constructor(
       ?.accumulateUntil(MIN_BUFFER_SIZE)
       ?.onEach { sendMediaStream(listOf(MediaData(it, "audio/pcm"))) }
       ?.catch { throw FirebaseAIException.from(it) }
-      ?.launchIn(CoroutineScope(Dispatchers.IO))
+      ?.launchIn(scope)
   }
 
   /**
@@ -338,7 +338,6 @@ internal constructor(
             } else if (functionCallHandler != null) {
               // It's fine to suspend here since you can't have a function call running concurrently
               // with an audio response
-              println("Model is attempting to send a function call response")
               sendFunctionResponse(it.functionCalls.map(functionCallHandler).toList())
             } else {
               Log.w(
@@ -354,13 +353,11 @@ internal constructor(
             )
           }
           is LiveServerContent -> {
-            println("State of it's interruption: ${it.interrupted}")
             if (it.interrupted) {
               playBackQueue.clear()
             } else {
               val audioParts = it.content?.parts?.filterIsInstance<InlineDataPart>().orEmpty()
               for (part in audioParts) {
-                println("Model receiving ${part.inlineData}")
                 playBackQueue.add(part.inlineData)
               }
             }
@@ -404,7 +401,6 @@ internal constructor(
           if (enableInterruptions != true) {
             audioHelper?.pauseRecording()
           }
-          println("Model playing $playbackData")
           audioHelper?.playAudio(playbackData)
         }
       }